+++ CHANGELOG.md
@@ -2,7 +2,11 @@ TBD
 ===
 Unreleased changes. Release notes have not yet been written.
 
-* separates crates: vu64, rabuf
+0.1.4 (2021-11-11)
+=====
+
+* renames FileDbMap to FileDbMapString, FileDbList to FileDbMapU64.
+* separates crates: vu64, rabuf.
 
 0.1.3 (2021-11-09)
 =====